Fix bug of combined alias APIs.

+++ server/controllers/AuthorController.js
@@ -635,18 +635,22 @@ class AuthorController {
   }
 
   /**
-   * DELETE: api/authors/combined_alias
+   * DELETE: api/authors/:id/combined_alias
    *
    * @param {import('express').Request} req
    * @param {import('express').Response} res
    */
   async deleteCombinedAlias(req, res) {
+    if (!req.body) {
+      return res.status(400).send('Missing request body')
+    }
+
     const { authorId, aliasId } = req.body
     try {
       const deleteResult = await Database.authorCombinedAliasModel.destroy({
         where: {
-          authorId,
-          aliasId
+          authorId: authorId,
+          aliasId: aliasId
         }
       })
 
@@ -655,13 +659,13 @@ class AuthorController {
       }
 
       const remainingAliases = await Database.authorCombinedAliasModel.count({
-        where:{ aliasId }
+        where:{ aliasId: aliasId }
       })
 
-      if (remainingAliases.length === 0) {
+      if (remainingAliases === 0) {
         await Database.authorModel.update(
           { is_alias_of: null },
-          { where: { id: aliasId } }
+          { where: {id: aliasId } }
         )
       }
